PascalLexer: Properly differentiate between overflow and other error kinds.
The i64::from_str() function returns different kinds of errors, namely overflow/underflow, invalid characters, etc. Unfortunately these error kinds are private, so we can't directly match against them. This workaround allows us to differentiate between overflow errors (which are valid errors that need to be handled properly) and other errors (which indicate a bug, so a panic is appropriate).